+++ packages/client/src/impl/node/NodeHttpTransport.ts
@@ -73,6 +73,8 @@ export class NodeHttpTransport implements Transport {
       port: url.port,
       protocol: url.protocol,
       hostname: url.hostname,
+      timeout:
+        nodeSupportedOptions.timeout ?? nodeSupportedOptions.writeTimeout,
     }
     this._contextPath = proxyUrl ? _url : url.path ?? ''
     if (this._contextPath.endsWith('/')) {
@@ -259,11 +261,15 @@ export class NodeHttpTransport implements Transport {
       yield chunk
     }
   }
+
   /**
    * Creates configuration for a specific request.
    *
    * @param path - API path starting with '/' and containing also query parameters
-   * @param body - request body, will be utf-8 encoded
+   * @param body - Request body, will be utf-8 encoded
+   * @param sendOptions - Options for sending a request message
+   * @param resolve - Resolve callback
+   * @param reject - Reject callback
    * @returns a configuration object that is suitable for making the request
    */
   private _createRequestMessage(
@@ -282,6 +288,7 @@ export class NodeHttpTransport implements Transport {
       const authScheme = this._authScheme ?? 'Token'
       headers.authorization = `${authScheme} ${this._token}`
     }
+
     const options: {[key: string]: any} = {
       ...this._defaultOptions,
       path: this._contextPath + path,
@@ -290,6 +297,7 @@ export class NodeHttpTransport implements Transport {
         ...headers,
         ...sendOptions.headers,
       },
+      timeout: sendOptions.timeout ?? this._defaultOptions.timeout,
     }
     if (sendOptions.signal) {
       options.signal = sendOptions.signal